+++ src/biglist/__init__.py
@@ -38,18 +38,18 @@ from ._parquet import (
     ParquetBiglist,
     ParquetFileReader,
     ParquetFileSeq,
+    make_parquet_field,
+    make_parquet_schema,
+    make_parquet_type,
     read_parquet_file,
+    write_arrays_to_parquet,
+    write_parquet_file,
+    write_pylist_to_parquet,
 )
 from ._util import (
     Chain,
     Seq,
     Slicer,
-    make_parquet_field,
-    make_parquet_schema,
-    make_parquet_type,
-    write_arrays_to_parquet,
-    write_parquet_file,
-    write_pylist_to_parquet,
 )
 
-__version__ = "0.7.8"
+__version__ = "0.7.9b1"