+++ tests/test_parquet.py
@@ -2,10 +2,21 @@ import io
 import random
 from types import SimpleNamespace
 from uuid import uuid4
+
 import pyarrow
-from upathlib import LocalUpath
-from biglist import ParquetBiglist, ParquetFileReader, write_arrays_to_parquet, read_parquet_file, Slicer, ParquetBatchData
 import pytest
+from biglist import (
+    ParquetBatchData,
+    ParquetBiglist,
+    ParquetFileReader,
+    Slicer,
+    make_parquet_schema,
+    make_parquet_type,
+    read_parquet_file,
+    write_arrays_to_parquet,
+    write_pylist_to_parquet,
+)
+from upathlib import LocalUpath
 
 
 def test_idx_locator():
@@ -25,16 +36,14 @@ def test_idx_locator():
     class My:
         def __init__(self):
             self.num_row_groups = 5
-            self.metadata = SimpleNamespace(
-                row_group=row_group_sizes
-            )
+            self.metadata = SimpleNamespace(row_group=row_group_sizes)
             self._row_groups_num_rows = None
             self._row_groups_num_rows_cumsum = None
             self._getitem_last_row_group = None
 
         def __getitem__(self, idx):
             return ParquetFileReader._locate_row_group_for_item(self, idx)
-        
+
     me = My()
     assert me[0] == (0, 0)
     assert me[1] == (0, 1)
@@ -51,7 +60,7 @@ def test_idx_locator():
     assert me[12] == (3, 0)
     assert me[13] == (4, 0)
     assert me[14] == (4, 1)
-    
+
     # jump around
     assert me[2] == (0, 2)
     assert me[11] == (2, 4)
@@ -130,7 +139,7 @@ def test_parquet_biglist(tmp_path):
             [str(uuid4()) for _ in range(N)],
         ],
         path / 'data_1.parquet',
-        names=['key', 'value']
+        names=['key', 'value'],
     )
 
     # key = pyarrow.array([random.randint(0, 10000) for _ in range(N)])
@@ -141,18 +150,18 @@ def test_parquet_biglist(tmp_path):
     # parquet.write_table(tab, str(path / 'd2' / 'data_2.parquet'))
 
     write_arrays_to_parquet(
-        [ 
-         [random.randint(0, 10000) for _ in range(N)],
-         [str(uuid4()) for _ in range(N)],
-         ],
+        [
+            [random.randint(0, 10000) for _ in range(N)],
+            [str(uuid4()) for _ in range(N)],
+        ],
         path / 'd2' / 'data_2.parquet',
-        names=['key', 'value']
+        names=['key', 'value'],
     )
 
     biglist = ParquetBiglist.new(path)
     assert len(biglist) == N + N
     assert len(biglist.files) == 2
-    
+
     print('')
     print('datafiles')
     z = biglist.files.data_files_info
@@ -198,7 +207,7 @@ def test_parquet_biglist(tmp_path):
     assert isinstance(d2[2], str)
     print(d2[2])
     with pytest.raises(ValueError):
-        d3 = d2.columns(['key'])
+        d2.columns(['key'])
     print(Slicer(d.columns(['key']))[7:17].collect())
     print(list(Slicer(d)[:7]))
 
@@ -212,7 +221,7 @@ def test_parquet_biglist(tmp_path):
     print(z)
     assert isinstance(z['key'], pyarrow.Int64Scalar)
     assert isinstance(z['value'], pyarrow.StringScalar)
-    
+
     # The `pyarrow.Scalar` types compare unequal to Python native types
     assert z['key'] != z['key'].as_py()
     assert z['value'] != z['value'].as_py()
@@ -246,3 +255,77 @@ def test_parquet_biglist(tmp_path):
         print(row)
         if k > 3:
             break
+
+
+def test_parquet_schema():
+    type_spec = (
+        'struct',
+        [
+            ('name', 'string', False),
+            ('age', 'uint8', True),
+            (
+                'income',
+                ('struct', (('currency', 'string'), ('amount', 'uint64'))),
+                False,
+            ),
+        ],
+    )
+    s = make_parquet_type(type_spec)
+    assert type(s) is pyarrow.StructType
+    print(s)
+
+    type_spec = ('map_', 'string', ('list_', 'int64'), True)
+    s = make_parquet_type(type_spec)
+    assert type(s) is pyarrow.MapType
+    print(s)
+
+    type_spec = ('list_', 'int64')
+    s = make_parquet_type(type_spec)
+    assert type(s) is pyarrow.ListType
+    print(s)
+
+    type_spec = ('list_', ('time32', 's'), 5)
+    s = make_parquet_type(type_spec)
+    assert type(s) is pyarrow.FixedSizeListType
+    print(s)
+
+    schema_spec = [
+        ('name', 'string', False),
+        ('age', 'uint8', False),
+        (
+            'income',
+            ('struct', (('concurrency', 'string', True), ('amount', 'uint64'))),
+            True,
+        ),
+        ('hobbies', ('list_', 'string'), True),
+    ]
+    schema = make_parquet_schema(schema_spec)
+    assert type(schema) is pyarrow.Schema
+    print('')
+    print(schema)
+
+
+def test_write_parquet_file(tmp_path):
+    data = [
+        {'name': 'tom', 'age': 38, 'income': {'concurrency': 'YEN', 'amount': 10000}},
+        {
+            'name': 'jane',
+            'age': 38,
+            'income': {'amount': 250},
+            'hobbies': ['soccer', 'swim'],
+        },
+        {'age': 38, 'hobbies': ['tennis', 'baseball']},
+        {'name': 'john', 'age': 20, 'income': {}, 'hobbies': ['soccer', 'swim']},
+        {'name': 'paul', 'age': 38, 'income': {'amount': 200}, 'hobbies': ['run']},
+    ]
+    schema_spec = [
+        ['name', 'string', False],
+        ['age', 'uint64'],
+        ['income', ['struct', [['currency', 'string'], ['amount', 'float64', True]]]],
+        ['hobbies', ['list_', 'string']],
+    ]
+    pp = tmp_path / 'data.parquet'
+    write_pylist_to_parquet(data, pp, schema_spec=schema_spec)
+    f = read_parquet_file(pp)
+    for row in f:
+        print(row)