+++ tests/test_util.py
@@ -1,10 +1,16 @@
 from __future__ import annotations
-from collections.abc import Sequence, Iterable, Sized
+
+from collections.abc import Iterable, Sequence, Sized
+
 import pytest
-from biglist._util import Slicer, Chain, Seq, locate_idx_in_chunked_seq, make_parquet_schema, make_parquet_type, write_pylist_to_parquet
 from biglist._base import BiglistBase, FileReader, FileSeq
-from biglist._parquet import ParquetBatchData, read_parquet_file
-import pyarrow
+from biglist._parquet import ParquetBatchData
+from biglist._util import (
+    Chain,
+    Seq,
+    Slicer,
+    locate_idx_in_chunked_seq,
+)
 
 
 def test_locate_idx_in_chunked_seq():
@@ -132,7 +138,7 @@ def test_slicer():
     llv = lv[::2]
     assert llv[2] == 9
     assert list(llv) == [1, 5, 9]
-    
+
     x = list(range(20))
     z: Slicer[list[int]] = Slicer(x, [2, 3, 5, 6, 13])
     assert z[3] == 6
@@ -146,7 +152,7 @@ def test_chain():
     mylist3 = list(range(18, 32))
     mylist: Chain[list[int]] = Chain(mylist1, mylist2, mylist3)
     data = list(range(32))
-    
+
     assert len(mylist) == len(data)
     assert list(mylist) == data
     assert mylist[3] == data[3]
@@ -167,62 +173,3 @@ def test_chain():
 
     with pytest.raises(IndexError):
         _ = ch[2]
-
-
-def test_parquet_schema():
-    type_spec = ('struct', 
-                 [('name', 'string', False),
-                  ('age', 'uint8', True), 
-                  ('income', ('struct', (('currency', 'string'), ('amount', 'uint64'))), False),
-                 ]
-                )
-    s = make_parquet_type(type_spec)
-    assert type(s) is pyarrow.StructType
-    print(s)
-                 
-    type_spec = ('map_', 'string', ('list_', 'int64'), True)
-    s = make_parquet_type(type_spec)
-    assert type(s) is pyarrow.MapType
-    print(s)
-
-    type_spec = ('list_', 'int64')
-    s = make_parquet_type(type_spec)
-    assert type(s) is pyarrow.ListType
-    print(s)
-
-    type_spec = ('list_', ('time32', 's'), 5)
-    s = make_parquet_type(type_spec)
-    assert type(s) is pyarrow.FixedSizeListType
-    print(s)
-
-    schema_spec = [
-        ('name', 'string', False),
-        ('age', 'uint8', False),
-        ('income', ('struct', (('concurrency', 'string', True), ('amount', 'uint64'))), True),
-        ('hobbies', ('list_', 'string'), True),
-    ]
-    schema = make_parquet_schema(schema_spec)
-    assert type(schema) is pyarrow.Schema
-    print('')
-    print(schema)
-
-
-def test_write_parquet_file(tmp_path):
-    data = [
-        {'name': 'tom', 'age': 38, 'income': {'concurrency': 'YEN', 'amount': 10000}},
-        {'name': 'jane', 'age': 38, 'income': {'amount': 250}, 'hobbies': ['soccer', 'swim']},
-        {'age': 38, 'hobbies': ['tennis', 'baseball']},
-        {'name': 'john', 'age': 20, 'income': {}, 'hobbies': ['soccer', 'swim']},
-        {'name': 'paul', 'age': 38, 'income': {'amount': 200}, 'hobbies': ['run']},
-    ]
-    schema_spec = [
-                    ['name', 'string', False],
-                    ['age', 'uint64'],
-                    ['income', ['struct', [['currency', 'string'], ['amount', 'float64', True]]]],
-                    ['hobbies', ['list_', 'string']],
-                ]
-    pp = tmp_path / 'data.parquet'
-    write_pylist_to_parquet(data, pp, schema_spec=schema_spec)
-    f = read_parquet_file(pp)
-    for row in f:
-        print(row)